added openweathermap to its own module
