+++ app.py
@@ -1,6 +1,5 @@
 # coding: utf-8
 import json
-import requests
 import os
 from datetime import datetime, timezone, timedelta
 
@@ -9,6 +8,7 @@ import dateutil.tz as tz # Used for timezone stuff
 
 import vasttrafik
 import skolmaten
+import openweathermap
 # import creds
 
 app = Flask(__name__)
@@ -205,76 +205,7 @@ def calculate_minutes(departure):
     else:
         return f'Ca {countdown}'
 
-def get_temperature():
-    if temp_situation["updated"] == 0 or temp_situation["updated"] < datetime.now(tz.gettz("Europe/Stockholm")) - timedelta(minutes=10):
-        print("Updating temperatures")
-        openweathermap_key = os.environ["OWM_KEY"].strip()
-        api_call = "https://api.openweathermap.org/data/2.5/weather?q=Göteborg&APPID=" + openweathermap_key
-        r = requests.get(api_call)
-        r_json = json.loads(str(r.json()).replace("'", '"'))
-        temp_now = str(round((r_json["main"]["temp"]-272.15), 1)) + "°C"
-        wheater_now = getWeatherEmoji(r_json["weather"][0]["id"])
 
-        hourly_call = "https://api.openweathermap.org/data/2.5/forecast?q=Göteborg&APPID=" + openweathermap_key
-        hourly_r = requests.get(hourly_call)
-        hourly_json = json.loads(str(hourly_r.json()).replace("'", '"'))
-        temp0 = str(round((hourly_json["list"][0]["main"]["temp"]-272.15),1)) + "°C"
-        temp0time = datetime.fromtimestamp(hourly_json["list"][0]["dt"], tz.gettz("Europe/Stockholm")).strftime('%H:%M')
-        temp0wheather = getWeatherEmoji(hourly_json["list"][0]["weather"][0]["id"])
-        temp1 = str(round((hourly_json["list"][1]["main"]["temp"]-272.15),1)) + "°C"
-        temp1time = datetime.fromtimestamp(hourly_json["list"][1]["dt"], tz.gettz("Europe/Stockholm")).strftime('%H:%M')
-        temp1wheather = getWeatherEmoji(hourly_json["list"][1]["weather"][0]["id"])
-        temp2 = str(round((hourly_json["list"][2]["main"]["temp"]-272.15),1)) + "°C"
-        temp2time = datetime.fromtimestamp(hourly_json["list"][2]["dt"], tz.gettz("Europe/Stockholm")).strftime('%H:%M')
-        temp2wheather = getWeatherEmoji(hourly_json["list"][2]["weather"][0]["id"])
-        
-        temp_situation["updated"] = datetime.now(tz.gettz("Europe/Stockholm"))
-        
-        out = [temp_now, wheater_now, temp0, temp0time, temp0wheather, temp1, temp1time, temp1wheather, temp2, temp2time, temp2wheather]
-        
-        temp_situation["last_temp"] = out
-        return out
-    else:
-        return temp_situation["last_temp"]
-
-def getWeatherEmoji(weatherID):
-
-    weatherIDstr = str(weatherID)
-
-    # Openweathermap Weather codes and corressponding emojis
-    thunderstorm = "\U0001F4A8"    # Code: 200's, 900, 901, 902, 905
-    drizzle = "\U0001F4A7"         # Code: 300's
-    rain = "\U00002614\U0000FE0F"            # Code: 500's
-    snowflake = "\U00002744\U0000FE0F"       # Code: 600's snowflake
-    # snowman = "\U000026C4"       # Code: 600's snowman, 903, 906
-    atmosphere = "\U0001F301"      # Code: 700's foogy
-    clearSky = "\U00002600\U0000FE0F"        # Code: 800 clear sky
-    fewClouds = "\U000026C5\U0000FE0F"       # Code: 801 sun behind clouds
-    clouds = "\U00002601\U0000FE0F"          # Code: 802-803-804 clouds general
-    hot = "\U0001F525"             # Code: 904
-    defaultEmoji = "\U0001F300"    # default emojis
-    
-    if weatherIDstr[0] == '2' or weatherIDstr == '900' or weatherIDstr == '901' or weatherIDstr == '902' or weatherIDstr == '905':
-        return thunderstorm
-    elif weatherIDstr[0] == '3':
-        return drizzle
-    elif weatherIDstr[0] == '5':
-        return rain
-    elif weatherIDstr[0] == '6' or weatherIDstr == '903' or weatherIDstr == '906':
-        return snowflake # + ' ' + snowman
-    elif weatherIDstr[0] == '7':
-        return atmosphere
-    elif weatherIDstr == '800':
-        return clearSky
-    elif weatherIDstr == '801':
-        return fewClouds
-    elif weatherIDstr == '802' or weatherIDstr == '803' or weatherIDstr == '804':
-        return clouds
-    elif weatherIDstr == '904':
-        return hot
-    else:
-        return defaultEmoji
-    
 # -------- INIT  --------
 key = os.environ["VT_KEY"]
 secret = os.environ["VT_SECRET"]
@@ -300,11 +231,6 @@ situation = {
     "situations": get_trafficsituation()
 }
 
-temp_situation = {
-    "updated": 0,
-    "last_temp": 0
-}
-
 # ------- ROUTES --------
 
 @app.route("/")
@@ -327,7 +253,7 @@ def getinfo():
     cpdep = format_departures(deps[2])
     kdep = format_departures(deps[3])
     disruptions = get_disruptions()
-    temperature = get_temperature()
+    temperature = openweathermap.get_temperature()
     menu = skolmaten.get_menu()
 
     d = {
